Consolidate admin section tests into page-level tests
Each admin entity now has a single page test covering both the details and
all of its forms (one test block per form), instead of split page + section
tests. Deletes the section test files and drops a no-op "renders ... details"
assertion. Adds form coverage for chapter lessons/problems, course chapters,
and problemset problems, which previously had none.
