+++ judgels-client/src/routes/admin/archives/ArchivePage/ArchivePage.test.jsx
@@ -1,4 +1,6 @@
-import { act, render, screen, within } from '@testing-library/react';
+import { act, render, screen, waitFor, within } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+import nock from 'nock';
 
 import { setSession } from '../../../../modules/session';
 import { QueryClientProviderWrapper } from '../../../../test/QueryClientProviderWrapper';
@@ -38,7 +40,7 @@ describe('ArchivePage', () => {
     );
   };
 
-  test('renders archive details', async () => {
+  test('details', async () => {
     await renderComponent();
 
     await screen.findByText(/Archive 1/);
@@ -58,5 +60,46 @@ describe('ArchivePage', () => {
       ['Category', 'Category 1'],
       ['Description', 'Description 1'],
     ]);
+  });
+
+  test('general form', async () => {
+    await renderComponent();
+
+    const user = userEvent.setup();
+
+    await user.click(await screen.findByRole('button', { name: /edit/i }));
+
+    const slug = screen.getByRole('textbox', { name: /slug/i });
+    expect(slug).toHaveValue('archive-1');
+    await user.clear(slug);
+    await user.type(slug, 'new-archive');
+
+    const names = screen.getAllByRole('textbox', { name: /name/i });
+    expect(names[0]).toHaveValue('Archive 1');
+    await user.clear(names[0]);
+    await user.type(names[0], 'New Archive');
+
+    const category = screen.getByRole('textbox', { name: /category/i });
+    expect(category).toHaveValue('Category 1');
+    await user.clear(category);
+    await user.type(category, 'New Category');
+
+    const description = document.querySelector('textarea[name="description"]');
+    expect(description).toHaveValue('Description 1');
+    await user.clear(description);
+    await user.type(description, 'New Description');
+
+    nockApi()
+      .post('/archives/JIDARCHIVE1', {
+        slug: 'new-archive',
+        name: 'New Archive',
+        category: 'New Category',
+        description: 'New Description',
+      })
+      .reply(200);
+
+    await user.click(screen.getByRole('button', { name: /save/i }));
+
+    await waitFor(() => expect(nock.isDone()).toBe(true));
   });
 });
